chore: remove sunset prospecting sources
- move ProspectType enum to content-common shared lib
