+++ lambdas/prospect-api-bridge-lambda/src/index.spec.ts
@@ -44,7 +44,7 @@ describe('processor', () => {
         scheduled_surface_guid: 'NEW_TAB_EN_US',
         prospect_id: 'a598e9cc-5f8a-5062-aa16-dbca19a45134',
         url: 'https://science.nasa.gov/directorates/smd/astrophysics-division/how-nasa-chases-and-investigates-bright-cosmic-blips/',
-        prospect_source: 'CONSTRAINT_SCHEDULE',
+        prospect_source: 'TIMESPENT_MODELED',
         save_count: 438,
         predicted_topic: 'SCIENCE',
         rank: 1,
@@ -53,7 +53,7 @@ describe('processor', () => {
         scheduled_surface_guid: 'NEW_TAB_EN_US',
         prospect_id: '9cc1ca00-6216-5a12-895b-37c48245fba2',
         url: 'https://www.scientificamerican.com/article/how-long-does-it-really-take-to-form-a-habit/',
-        prospect_source: 'CONSTRAINT_SCHEDULE',
+        prospect_source: 'TIMESPENT_MODELED',
         save_count: 0,
         predicted_topic: 'HEALTH_FITNESS',
         rank: 2,
