+++ servers/curated-corpus-api/schema-admin.graphql
@@ -70,7 +70,7 @@ enum RejectionReason {
 2024-01-29
 
 these reasons are used when dismissing prospects only for the
-SLATE_SCHEDULER prospect type. this is a preliminary step towards ML
+SLATE_SCHEDULE_V2 prospect type. this is a preliminary step towards ML
 scheduling items, and this reason set will *likely* be used when removing ML
 scheduled items as well.
 
@@ -133,8 +133,6 @@ enum ProspectType {
   DISMISSED
   RSS_LOGISTIC
   RSS_LOGISTIC_RECENT
-  CONSTRAINT_SCHEDULE
-  SLATE_SCHEDULER
   SLATE_SCHEDULER_V2
 }
