+++ servers/curated-corpus-api/src/admin/resolvers/mutations/ScheduledItem.ts
@@ -78,8 +78,6 @@ export async function deleteScheduledItem(
     scheduledCorpusItem: {
       ...scheduledItem,
       status: ScheduledCorpusItemStatus.REMOVED,
-      // hard-coded to manual for now. once MC-645 is complete, this should be pulled from
-      // the `scheduledItem.source` property.
       generated_by: scheduledItem.source as ScheduledItemSource,
       // get reasons and reason comment (these may both be null. they're only
       // supplied when a scheduled item is deleted for a limited set of surfaces)
